Fix loading scene for every request on script file for workspace completion
For workspace completion the Scene for the script is loaded for every
completion request and freed afterwards. Resulting in performance issues
for scenes including large resources.

Loaded Scenes are now cached for each script and loaded when opening the
script. Not loading them again and only freeing them when the connection
is terminated, the language server stopped or the script is closed.